# apps/api/openapi.json
{
	"openapi": "3.1.1",
	"info": {
		"title": "Talqo API",
		"version": "0.0.1"
	},
	"components": {
		"securitySchemes": {
			"SessionCookie": {
				"in": "cookie",
				"name": "session",
				"type": "apiKey"
			}
		},
		"schemas": {
...
			"RoleUser": {
				"type": "object",
				"properties": {
					"id": {
						"type": "string"
					},
					"username": {
						"type": "string"
					},
					"mustChangePassword": {
						"type": "boolean"
					}
				},
				"required": ["id", "username", "mustChangePassword"]
			},
...
		},
...
	},
	"paths": {
...
		"/api/access": {
			"get": {
				"operationId": "getAccess",
				"tags": ["Roles"],
				"security": [
					{
						"SessionCookie": []
					}
				],
				"responses": {
					"200": {
						"description": "Effective access for the current user",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"permissions": {
											"type": "array",
											"items": {
												"type": "string",
												"enum": ["admin", "users:invite", "ai_provider:manage", "agents:read", "agents:manage"]
											}
										}
									},
									"required": ["permissions"]
								}
							}
						}
					},
					"400": {
						"description": "https://docs.talqo.chat/problems",
						"content": {
							"application/problem+json": {
								"schema": {
									"oneOf": [
										{
											"additionalProperties": false,
											"properties": {
												"code": {
													"const": "malformed-json",
													"type": "string"
												},
												"type": {
													"const": "https://docs.talqo.chat/problems#malformed-json",
													"type": "string"
												}
											},
											"required": ["code", "type"],
											"type": "object"
										}
									]
								}
							}
						}
					},
					"401": {
						"description": "https://docs.talqo.chat/problems",
						"content": {
							"application/problem+json": {
								"schema": {
									"oneOf": [
										{
											"additionalProperties": false,
											"properties": {
												"code": {
													"const": "authentication-required",
													"type": "string"
												},
												"type": {
													"const": "https://docs.talqo.chat/problems#authentication-required",
													"type": "string"
												}
											},
											"required": ["code", "type"],
											"type": "object"
										}
									]
								}
							}
						}
					},
					"403": {
						"description": "https://docs.talqo.chat/problems",
						"content": {
							"application/problem+json": {
								"schema": {
									"oneOf": [
										{
											"additionalProperties": false,
											"properties": {
												"code": {
													"const": "password-change-required",
													"type": "string"
												},
												"type": {
													"const": "https://docs.talqo.chat/problems#password-change-required",
													"type": "string"
												}
											},
											"required": ["code", "type"],
											"type": "object"
										}
									]
								}
							}
						}
					},
					"500": {
						"description": "https://docs.talqo.chat/problems",
						"content": {
							"application/problem+json": {
								"schema": {
									"oneOf": [
										{
											"additionalProperties": false,
											"properties": {
												"code": {
													"const": "internal-server-error",
													"type": "string"
												},
												"type": {
													"const": "https://docs.talqo.chat/problems#internal-server-error",
													"type": "string"
												}
											},
											"required": ["code", "type"],
											"type": "object"
										}
									]
								}
							}
						}
					}
				}
			}
		},
		"/api/setup": {
			"get": {
				"operationId": "getSetupStatus",
				"tags": ["Roles"],
				"responses": {
					"200": {
						"description": "Setup status",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"needsSetup": {
											"type": "boolean"
										}
									},
									"required": ["needsSetup"]
								}
							}
						}
					},
					"400": {
						"description": "https://docs.talqo.chat/problems",
						"content": {
							"application/problem+json": {
								"schema": {
									"oneOf": [
										{
											"additionalProperties": false,
											"properties": {
												"code": {
													"const": "malformed-json",
													"type": "string"
												},
												"type": {
													"const": "https://docs.talqo.chat/problems#malformed-json",
													"type": "string"
												}
											},
											"required": ["code", "type"],
											"type": "object"
										}
									]
								}
							}
						}
					},
					"500": {
						"description": "https://docs.talqo.chat/problems",
						"content": {
							"application/problem+json": {
								"schema": {
									"oneOf": [
										{
											"additionalProperties": false,
											"properties": {
												"code": {
													"const": "internal-server-error",
													"type": "string"
												},
												"type": {
													"const": "https://docs.talqo.chat/problems#internal-server-error",
													"type": "string"
												}
											},
											"required": ["code", "type"],
											"type": "object"
										}
									]
								}
							}
						}
					}
				}
			},
			"post": {
				"operationId": "bootstrapAdmin",
				"tags": ["Roles"],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"username": {
										"type": "string",
										"minLength": 3,
										"maxLength": 32,
										"pattern": "^[a-zA-Z0-9_-]+$"
									},
									"password": {
										"type": "string",
										"minLength": 8,
										"maxLength": 128
									}
								},
								"required": ["username", "password"]
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Admin created",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"user": {
											"$ref": "#/components/schemas/RoleUser"
										}
									},
									"required": ["user"]
								}
							}
						}
					},
					"400": {
						"description": "https://docs.talqo.chat/problems",
						"content": {
							"application/problem+json": {
								"schema": {
									"oneOf": [
										{
											"additionalProperties": false,
											"properties": {
												"code": {
													"const": "invalid-request",
													"type": "string"
												},
												"type": {
													"const": "https://docs.talqo.chat/problems#invalid-request",
													"type": "string"
												}
											},
											"required": ["code", "type"],
											"type": "object"
										},
										{
											"additionalProperties": false,
											"properties": {
												"code": {
													"const": "malformed-json",
													"type": "string"
												},
												"type": {
													"const": "https://docs.talqo.chat/problems#malformed-json",
													"type": "string"
												}
											},
											"required": ["code", "type"],
											"type": "object"
										}
									]
								}
							}
						}
					},
					"409": {
						"description": "https://docs.talqo.chat/problems",
						"content": {
							"application/problem+json": {
								"schema": {
									"oneOf": [
										{
											"additionalProperties": false,
											"properties": {
												"code": {
													"const": "admin-already-exists",
													"type": "string"
												},
												"type": {
													"const": "https://docs.talqo.chat/problems#admin-already-exists",
													"type": "string"
												}
											},
											"required": ["code", "type"],
											"type": "object"
										},
										{
											"additionalProperties": false,
											"properties": {
												"code": {
													"const": "username-taken",
													"type": "string"
												},
												"type": {
													"const": "https://docs.talqo.chat/problems#username-taken",
													"type": "string"
												}
											},
											"required": ["code", "type"],
											"type": "object"
										}
									]
								}
							}
						}
					},
					"500": {
						"description": "https://docs.talqo.chat/problems",
						"content": {
							"application/problem+json": {
								"schema": {
									"oneOf": [
										{
											"additionalProperties": false,
											"properties": {
												"code": {
													"const": "internal-server-error",
													"type": "string"
												},
												"type": {
													"const": "https://docs.talqo.chat/problems#internal-server-error",
													"type": "string"
												}
											},
											"required": ["code", "type"],
											"type": "object"
										}
									]
								}
							}
						}
					}
				}
			}
		},
		"/api/invitations": {
			"post": {
				"operationId": "createInvitation",
				"tags": ["Roles"],
				"security": [
					{
						"SessionCookie": []
					}
				],
				"responses": {
					"201": {
						"description": "Invitation created",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"token": {
											"type": "string"
										},
										"expiresAt": {
											"type": "string",
											"format": "date-time"
										}
									},
									"required": ["token", "expiresAt"]
								}
							}
						}
					},
					"400": {
						"description": "https://docs.talqo.chat/problems",
						"content": {
							"application/problem+json": {
								"schema": {
									"oneOf": [
										{
											"additionalProperties": false,
											"properties": {
												"code": {
													"const": "malformed-json",
													"type": "string"
												},
												"type": {
													"const": "https://docs.talqo.chat/problems#malformed-json",
													"type": "string"
												}
											},
											"required": ["code", "type"],
											"type": "object"
										}
									]
								}
							}
						}
					},
					"401": {
						"description": "https://docs.talqo.chat/problems",
						"content": {
							"application/problem+json": {
								"schema": {
									"oneOf": [
										{
											"additionalProperties": false,
											"properties": {
												"code": {
													"const": "authentication-required",
													"type": "string"
												},
												"type": {
													"const": "https://docs.talqo.chat/problems#authentication-required",
													"type": "string"
												}
											},
											"required": ["code", "type"],
											"type": "object"
										}
									]
								}
							}
						}
					},
					"403": {
						"description": "https://docs.talqo.chat/problems",
						"content": {
							"application/problem+json": {
								"schema": {
									"oneOf": [
										{
											"additionalProperties": false,
											"properties": {
												"code": {
													"const": "password-change-required",
													"type": "string"
												},
												"type": {
													"const": "https://docs.talqo.chat/problems#password-change-required",
													"type": "string"
												}
											},
											"required": ["code", "type"],
											"type": "object"
										},
										{
											"additionalProperties": false,
											"properties": {
												"code": {
													"const": "permission-denied",
													"type": "string"
												},
												"type": {
													"const": "https://docs.talqo.chat/problems#permission-denied",
													"type": "string"
												}
											},
											"required": ["code", "type"],
											"type": "object"
										}
									]
								}
							}
						}
					},
					"500": {
						"description": "https://docs.talqo.chat/problems",
						"content": {
							"application/problem+json": {
								"schema": {
									"oneOf": [
										{
											"additionalProperties": false,
											"properties": {
												"code": {
													"const": "internal-server-error",
													"type": "string"
												},
												"type": {
													"const": "https://docs.talqo.chat/problems#internal-server-error",
													"type": "string"
												}
											},
											"required": ["code", "type"],
											"type": "object"
										}
									]
								}
							}
						}
					}
				}
			}
		},
		"/api/invitations/redeem": {
			"post": {
				"operationId": "redeemInvitation",
				"tags": ["Roles"],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"token": {
										"type": "string",
										"minLength": 1,
										"maxLength": 128
									},
									"username": {
										"type": "string",
										"minLength": 3,
										"maxLength": 32,
										"pattern": "^[a-zA-Z0-9_-]+$"
									},
									"password": {
										"type": "string",
										"minLength": 8,
										"maxLength": 128
									}
								},
								"required": ["token", "username", "password"]
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Invitation redeemed",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"user": {
											"$ref": "#/components/schemas/RoleUser"
										}
									},
									"required": ["user"]
								}
							}
						}
					},
					"400": {
						"description": "https://docs.talqo.chat/problems",
						"content": {
							"application/problem+json": {
								"schema": {
									"oneOf": [
										{
											"additionalProperties": false,
											"properties": {
												"code": {
													"const": "invalid-request",
													"type": "string"
												},
												"type": {
													"const": "https://docs.talqo.chat/problems#invalid-request",
													"type": "string"
												}
											},
											"required": ["code", "type"],
											"type": "object"
										},
										{
											"additionalProperties": false,
											"properties": {
												"code": {
													"const": "malformed-json",
													"type": "string"
												},
												"type": {
													"const": "https://docs.talqo.chat/problems#malformed-json",
													"type": "string"
												}
											},
											"required": ["code", "type"],
											"type": "object"
										}
									]
								}
							}
						}
					},
					"409": {
						"description": "https://docs.talqo.chat/problems",
						"content": {
							"application/problem+json": {
								"schema": {
									"oneOf": [
										{
											"additionalProperties": false,
											"properties": {
												"code": {
													"const": "invalid-invitation",
													"type": "string"
												},
												"type": {
													"const": "https://docs.talqo.chat/problems#invalid-invitation",
													"type": "string"
												}
											},
											"required": ["code", "type"],
											"type": "object"
										},
										{
											"additionalProperties": false,
											"properties": {
												"code": {
													"const": "username-taken",
													"type": "string"
												},
												"type": {
													"const": "https://docs.talqo.chat/problems#username-taken",
													"type": "string"
												}
											},
											"required": ["code", "type"],
											"type": "object"
										}
									]
								}
							}
						}
					},
					"500": {
						"description": "https://docs.talqo.chat/problems",
						"content": {
							"application/problem+json": {
								"schema": {
									"oneOf": [
										{
											"additionalProperties": false,
											"properties": {
												"code": {
													"const": "internal-server-error",
													"type": "string"
												},
												"type": {
													"const": "https://docs.talqo.chat/problems#internal-server-error",
													"type": "string"
												}
											},
											"required": ["code", "type"],
											"type": "object"
										}
									]
								}
							}
						}
					}
				}
			}
		},
		"/api/permission-grants": {
			"post": {
				"operationId": "createPermissionGrant",
				"tags": ["Roles"],
				"security": [
					{
						"SessionCookie": []
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"userId": {
										"type": "string",
										"minLength": 1,
										"maxLength": 128
									},
									"permission": {
										"type": "string",
										"enum": ["admin", "users:invite", "ai_provider:manage", "agents:read", "agents:manage"]
									}
								},
								"required": ["userId", "permission"],
								"additionalProperties": false
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Permission granted",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"grant": {
											"type": "object",
											"properties": {
												"id": {
													"type": "string"
												},
												"userId": {
													"type": "string"
												},
												"permission": {
													"type": "string"
												},
												"grantedBy": {
													"type": ["string", "null"]
												},
												"grantedAt": {
													"type": "string",
													"format": "date-time"
												}
											},
											"required": ["id", "userId", "permission", "grantedBy", "grantedAt"]
										}
									},
									"required": ["grant"]
								}
							}
						}
					},
					"400": {
						"description": "https://docs.talqo.chat/problems",
						"content": {
							"application/problem+json": {
								"schema": {
									"oneOf": [
										{
											"additionalProperties": false,
											"properties": {
												"code": {
													"const": "invalid-request",
													"type": "string"
												},
												"type": {
													"const": "https://docs.talqo.chat/problems#invalid-request",
													"type": "string"
												}
											},
											"required": ["code", "type"],
											"type": "object"
										},
										{
											"additionalProperties": false,
											"properties": {
												"code": {
													"const": "malformed-json",
													"type": "string"
												},
												"type": {
													"const": "https://docs.talqo.chat/problems#malformed-json",
													"type": "string"
												}
											},
											"required": ["code", "type"],
											"type": "object"
										}
									]
								}
							}
						}
					},
					"401": {
						"description": "https://docs.talqo.chat/problems",
						"content": {
							"application/problem+json": {
								"schema": {
									"oneOf": [
										{
											"additionalProperties": false,
											"properties": {
												"code": {
													"const": "authentication-required",
													"type": "string"
												},
												"type": {
													"const": "https://docs.talqo.chat/problems#authentication-required",
													"type": "string"
												}
											},
											"required": ["code", "type"],
											"type": "object"
										}
									]
								}
							}
						}
					},
					"403": {
						"description": "https://docs.talqo.chat/problems",
						"content": {
							"application/problem+json": {
								"schema": {
									"oneOf": [
										{
											"additionalProperties": false,
											"properties": {
												"code": {
													"const": "admin-access-required",
													"type": "string"
												},
												"type": {
													"const": "https://docs.talqo.chat/problems#admin-access-required",
													"type": "string"
												}
											},
											"required": ["code", "type"],
											"type": "object"
										},
										{
											"additionalProperties": false,
											"properties": {
												"code": {
													"const": "password-change-required",
													"type": "string"
												},
												"type": {
													"const": "https://docs.talqo.chat/problems#password-change-required",
													"type": "string"
												}
											},
											"required": ["code", "type"],
											"type": "object"
										}
									]
								}
							}
						}
					},
					"404": {
						"description": "https://docs.talqo.chat/problems",
						"content": {
							"application/problem+json": {
								"schema": {
									"oneOf": [
										{
											"additionalProperties": false,
											"properties": {
												"code": {
													"const": "user-not-found",
													"type": "string"
												},
												"type": {
													"const": "https://docs.talqo.chat/problems#user-not-found",
													"type": "string"
												}
											},
											"required": ["code", "type"],
											"type": "object"
										}
									]
								}
							}
						}
					},
					"409": {
						"description": "https://docs.talqo.chat/problems",
						"content": {
							"application/problem+json": {
								"schema": {
									"oneOf": [
										{
											"additionalProperties": false,
											"properties": {
												"code": {
													"const": "admin-already-exists",
													"type": "string"
												},
												"type": {
													"const": "https://docs.talqo.chat/problems#admin-already-exists",
													"type": "string"
												}
											},
											"required": ["code", "type"],
											"type": "object"
										}
									]
								}
							}
						}
					},
					"500": {
						"description": "https://docs.talqo.chat/problems",
						"content": {
							"application/problem+json": {
								"schema": {
									"oneOf": [
										{
											"additionalProperties": false,
											"properties": {
												"code": {
													"const": "internal-server-error",
													"type": "string"
												},
												"type": {
													"const": "https://docs.talqo.chat/problems#internal-server-error",
													"type": "string"
												}
											},
											"required": ["code", "type"],
											"type": "object"
										}
									]
								}
							}
						}
					}
				}
			}
		},
...
		"/api/users": {
			"get": {
				"operationId": "listUsers",
				"tags": ["Roles"],
				"security": [
					{
						"SessionCookie": []
					}
				],
				"responses": {
					"200": {
						"description": "All users",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"properties": {
										"users": {
											"type": "array",
											"items": {
												"$ref": "#/components/schemas/RoleUser"
											}
										}
									},
									"required": ["users"]
								}
							}
						}
					},
					"400": {
						"description": "https://docs.talqo.chat/problems",
						"content": {
							"application/problem+json": {
								"schema": {
									"oneOf": [
										{
											"additionalProperties": false,
											"properties": {
												"code": {
													"const": "malformed-json",
													"type": "string"
												},
												"type": {
													"const": "https://docs.talqo.chat/problems#malformed-json",
													"type": "string"
												}
											},
											"required": ["code", "type"],
											"type": "object"
										}
									]
								}
							}
						}
					},
					"401": {
						"description": "https://docs.talqo.chat/problems",
						"content": {
							"application/problem+json": {
								"schema": {
									"oneOf": [
										{
											"additionalProperties": false,
											"properties": {
												"code": {
													"const": "authentication-required",
													"type": "string"
												},
												"type": {
													"const": "https://docs.talqo.chat/problems#authentication-required",
													"type": "string"
												}
											},
											"required": ["code", "type"],
											"type": "object"
										}
									]
								}
							}
						}
					},
					"403": {
						"description": "https://docs.talqo.chat/problems",
						"content": {
							"application/problem+json": {
								"schema": {
									"oneOf": [
										{
											"additionalProperties": false,
											"properties": {
												"code": {
													"const": "admin-access-required",
													"type": "string"
												},
												"type": {
													"const": "https://docs.talqo.chat/problems#admin-access-required",
													"type": "string"
												}
											},
											"required": ["code", "type"],
											"type": "object"
										},
										{
											"additionalProperties": false,
											"properties": {
												"code": {
													"const": "password-change-required",
													"type": "string"
												},
												"type": {
													"const": "https://docs.talqo.chat/problems#password-change-required",
													"type": "string"
												}
											},
											"required": ["code", "type"],
											"type": "object"
										}
									]
								}
							}
						}
					},
					"500": {
						"description": "https://docs.talqo.chat/problems",
						"content": {
							"application/problem+json": {
								"schema": {
									"oneOf": [
										{
											"additionalProperties": false,
											"properties": {
												"code": {
													"const": "internal-server-error",
													"type": "string"
												},
												"type": {
													"const": "https://docs.talqo.chat/problems#internal-server-error",
													"type": "string"
												}
											},
											"required": ["code", "type"],
											"type": "object"
										}
									]
								}
							}
						}
					}
				}
			}
		},
...
	},
...
}